backend/requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
requests>=2.28.0
PyJWT>=2.8.0
redis>=5.0.0
shapely>=2.0.0
zeep>=4.2.0
pyproj>=3.6.0

